adjusted Collider on both player. Added a system to check frame per frame in animations. Added a framechecker and a animationclipinfo script. Added an animation extension for the new frame system. Adjusted the attack script to store all relevant data and to be adjustable inside the inspector

+++ Assets/+++Workdata/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    public void Damage(int damageAmount, float stunDuration, float hitStopDuration);
+    public void Damage(float damageAmount, float stunDuration, float hitStopDuration);
 }
